Change validator range from long type to int type

+++ src/main/java/com/github/dahaka934/jhocon/annotations/ValidatorRange.java
@@ -19,7 +19,7 @@ import java.lang.annotation.*;
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.FIELD)
 public @interface ValidatorRange {
-    long min() default Long.MIN_VALUE;
+    int min() default Integer.MIN_VALUE;
 
-    long max() default Long.MAX_VALUE;
+    int max() default Integer.MAX_VALUE;
 }
